+++ src/DB.php
@@ -493,15 +493,16 @@ class DB
         try {
             $sth = self::$pdo->prepare('
                 INSERT IGNORE INTO `' . TB_CHAT . '`
-                (`id`, `type`, `title`, `username`, `first_name`, `last_name`, `created_at` ,`updated_at`, `old_id`)
+                (`id`, `type`, `title`, `username`, `first_name`, `last_name`, `is_forum`, `created_at` ,`updated_at`, `old_id`)
                 VALUES
-                (:id, :type, :title, :username, :first_name, :last_name, :created_at, :updated_at, :old_id)
+                (:id, :type, :title, :username, :first_name, :last_name, :is_forum, :created_at, :updated_at, :old_id)
                 ON DUPLICATE KEY UPDATE
                     `type`                           = VALUES(`type`),
                     `title`                          = VALUES(`title`),
                     `username`                       = VALUES(`username`),
                     `first_name`                     = VALUES(`first_name`),
                     `last_name`                      = VALUES(`last_name`),
+                    `is_forum`                       = VALUES(`is_forum`),
                     `updated_at`                     = VALUES(`updated_at`)
             ');
 
@@ -523,6 +524,7 @@ class DB
             $sth->bindValue(':username', $chat->getUsername());
             $sth->bindValue(':first_name', $chat->getFirstName());
             $sth->bindValue(':last_name', $chat->getLastName());
+            $sth->bindValue(':is_forum', $chat->getIsForum());
             $date = $date ?: self::getTimestamp();
             $sth->bindValue(':created_at', $date);
             $sth->bindValue(':updated_at', $date);
@@ -1131,24 +1133,26 @@ class DB
             $sth = self::$pdo->prepare('
                 INSERT IGNORE INTO `' . TB_MESSAGE . '`
                 (
-                    `id`, `user_id`, `chat_id`, `sender_chat_id`, `date`, `forward_from`, `forward_from_chat`, `forward_from_message_id`,
-                    `forward_signature`, `forward_sender_name`, `forward_date`,
+                    `id`, `user_id`, `chat_id`, `message_thread_id`, `sender_chat_id`, `date`, `forward_from`, `forward_from_chat`, `forward_from_message_id`,
+                    `forward_signature`, `forward_sender_name`, `forward_date`, `is_topic_message`,
                     `reply_to_chat`, `reply_to_message`, `via_bot`, `edit_date`, `media_group_id`, `author_signature`, `text`, `entities`, `caption_entities`,
                     `audio`, `document`, `animation`, `game`, `photo`, `sticker`, `video`, `voice`, `video_note`, `caption`, `contact`,
                     `location`, `venue`, `poll`, `dice`, `new_chat_members`, `left_chat_member`,
                     `new_chat_title`, `new_chat_photo`, `delete_chat_photo`, `group_chat_created`,
                     `supergroup_chat_created`, `channel_chat_created`, `message_auto_delete_timer_changed`, `migrate_to_chat_id`, `migrate_from_chat_id`,
                     `pinned_message`, `invoice`, `successful_payment`, `connected_website`, `passport_data`, `proximity_alert_triggered`,
+                    `forum_topic_created`, `forum_topic_closed`, `forum_topic_reopened`,
                     `video_chat_scheduled`, `video_chat_started`, `video_chat_ended`, `video_chat_participants_invited`, `web_app_data`, `reply_markup`
                 ) VALUES (
-                    :message_id, :user_id, :chat_id, :sender_chat_id, :date, :forward_from, :forward_from_chat, :forward_from_message_id,
-                    :forward_signature, :forward_sender_name, :forward_date,
+                    :message_id, :user_id, :chat_id, :message_thread_id, :sender_chat_id, :date, :forward_from, :forward_from_chat, :forward_from_message_id,
+                    :forward_signature, :forward_sender_name, :forward_date, :is_topic_message,
                     :reply_to_chat, :reply_to_message, :via_bot, :edit_date, :media_group_id, :author_signature, :text, :entities, :caption_entities,
                     :audio, :document, :animation, :game, :photo, :sticker, :video, :voice, :video_note, :caption, :contact,
                     :location, :venue, :poll, :dice, :new_chat_members, :left_chat_member,
                     :new_chat_title, :new_chat_photo, :delete_chat_photo, :group_chat_created,
                     :supergroup_chat_created, :channel_chat_created, :message_auto_delete_timer_changed, :migrate_to_chat_id, :migrate_from_chat_id,
                     :pinned_message, :invoice, :successful_payment, :connected_website, :passport_data, :proximity_alert_triggered,
+                    :forum_topic_created, :forum_topic_closed, :forum_topic_reopened,
                     :video_chat_scheduled, :video_chat_started, :video_chat_ended, :video_chat_participants_invited, :web_app_data, :reply_markup
                 )
             ');
@@ -1167,6 +1171,7 @@ class DB
             $sth->bindValue(':message_id', $message->getMessageId());
             $sth->bindValue(':chat_id', $chat_id);
             $sth->bindValue(':sender_chat_id', $sender_chat_id);
+            $sth->bindValue(':message_thread_id', $message->getMessageThreadId());
             $sth->bindValue(':user_id', $user_id);
             $sth->bindValue(':date', $date);
             $sth->bindValue(':forward_from', $forward_from);
@@ -1175,6 +1180,7 @@ class DB
             $sth->bindValue(':forward_signature', $message->getForwardSignature());
             $sth->bindValue(':forward_sender_name', $message->getForwardSenderName());
             $sth->bindValue(':forward_date', $forward_date);
+            $sth->bindValue(':is_topic_message', $message->getIsTopicMessage());
 
             $reply_to_chat_id = null;
             if ($reply_to_message_id !== null) {
@@ -1222,6 +1228,9 @@ class DB
             $sth->bindValue(':connected_website', $message->getConnectedWebsite());
             $sth->bindValue(':passport_data', $message->getPassportData());
             $sth->bindValue(':proximity_alert_triggered', $message->getProximityAlertTriggered());
+            $sth->bindValue(':forum_topic_created', $message->getForumTopicCreated());
+            $sth->bindValue(':forum_topic_closed', $message->getForumTopicClosed());
+            $sth->bindValue(':forum_topic_reopened', $message->getForumTopicReopened());
             $sth->bindValue(':video_chat_scheduled', $message->getVideoChatScheduled());
             $sth->bindValue(':video_chat_started', $message->getVideoChatStarted());
             $sth->bindValue(':video_chat_ended', $message->getVideoChatEnded());