+++ src/Entities/Topics/ForumTopicClosed.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Longman\TelegramBot\Entities\Topics;
+
+/**
+ * Class ForumTopicClosed
+ *
+ * This object represents a service message about a forum topic closed in the chat. Currently holds no information.
+ *
+ * @link https://core.telegram.org/bots/api#forumtopicclosed
+ */
+class ForumTopicClosed extends \Longman\TelegramBot\Entities\Entity
+{
+
+}